+++ CHANGELOG.md
@@ -1,12 +1,36 @@
 # Change Log
 
+## Release - 1.1.0
+
+- 2nd Public release.
+
+### Added in 1.1.0
+
+- Bracket Pair Guide colors.
+- A few of Git colors.
+
+### Updated in 1.1.0
+
+- UI's Background and Foreground colors.
+- Symbol Icon colors.
+- Debug Icon colors.
+- Fixed Input Control related colors.
+- Fixed List Hover colors.
+- Styx and Styx+ Syntax Highligthing colors.
+
+### Removed in 1.1.0
+
+- None.
+
+</br>
+
 ## Release - 1.0.0
 
-- Initial release
+- Initial release.
 
-### Added
+### Added in 1.0.0
 
 - Themed UI.
-- 7 Syntax Highlighting Schemes (Styx, Styx+, Styx Dracula, Styx Fleet, Styx Material, Styx Monokai, Styx VSCode)
+- 7 Syntax Highlighting Schemes (Styx, Styx+, Styx Dracula, Styx Fleet, Styx Material, Styx Monokai, Styx VSCode).
 - Syntax Highlighting support for Python, C, C++, Julia, Java, Kotlin (Partial), JavaScript and TypeScript.
 - Syntax Highlighting support for HTML, CSS and Markdown.